pyproject.toml
[project]
name = "sweetbalance-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "streamlit>=1.56.0",
    "streamlit-shadcn-ui>=0.1.19",
]
